+++ packages/transaction/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.29](https://github.com/rarible/sdk/compare/v0.8.28...v0.8.29) (2022-02-04)
+
+**Note:** Version bump only for package @rarible/sdk-transaction
+
+
+
+
+
+
 ## [0.8.28](https://github.com/rarible/sdk/compare/v0.8.27...v0.8.28) (2022-02-03)